+++ conda/plugins/subcommands/__init__.py
@@ -1,5 +1,5 @@
 # Copyright (C) 2012 Anaconda, Inc
 # SPDX-License-Identifier: BSD-3-Clause
-from .doctor import cli as doctor
+from . import doctor
 
 plugins = [doctor]
